Look into handling the the default `danger` command better
