+++ source/commands/danger.ts
@@ -1,9 +1,12 @@
 #! /usr/bin/env node
 
-import { version } from "../../package.json"
 import * as program from "commander"
 import * as debug from "debug"
 import chalk from "chalk"
+
+import { version } from "../../package.json"
+import setSharedArgs, { SharedCLI } from "./utils/sharedDangerfileArgs"
+import { runRunner } from "./run/runner"
 
 const d = debug("danger:runner")
 d(`argv: ${process.argv}`)
@@ -14,11 +17,16 @@ process.on("unhandledRejection", function(reason: string, _p: any) {
 })
 
 // Provides the root node to the command-line architecture
+setSharedArgs(program)
+
 program
   .version(version)
   .command("init", "Helps you get started with Danger")
   .command("process", "Like `run` but lets another process handle evaluating a Dangerfile")
   .command("pr", "Runs your changes against an existing PR")
   .command("runner", "Runs a dangerfile against a DSL passed in via STDIN")
-  .command("run", "Runs danger on your local system", { isDefault: true })
+  .command("run", "Runs danger on your local system")
   .parse(process.argv)
+
+const app = (program as any) as SharedCLI
+runRunner(app)
